# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación para API REST de Recetas",
        "contact": {
            "email": "eescobar@cdhidalgo.tecnm.mx"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apibecas2-production.up.railway.app"
        }
    ],
    "paths": {
...
        "/api/social": {
            "get": {
                "tags": [
                    "Social"
                ],
                "summary": "Obtener lista de registros sociales",
                "operationId": "3ec3b0c1753f88c38733926761ef6a63",
                "responses": {
                    "200": {
                        "description": "Lista de registros sociales recuperada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "id_estudiante": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "integrantes_familia": {
                                                "type": "integer",
                                                "example": 4
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado - Requiere permiso 'Modificar registros'"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Social"
                ],
                "summary": "Crear un nuevo registro social",
                "operationId": "c61deee436b3e370728de17c4febcfa5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_estudiante": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "integrantes_familia": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registro social creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "id_estudiante": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "integrantes_familia": {
                                            "type": "integer",
                                            "example": 4
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado - Requiere permiso 'CrearEliminar'"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/social/{id}": {
            "get": {
                "tags": [
                    "Social"
                ],
                "summary": "Obtener un registro social específico",
                "operationId": "14ca4341051d96270ef5a6fcb3dcd2e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del registro social",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro social encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "id_estudiante": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "integrantes_familia": {
                                            "type": "integer",
                                            "example": 4
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro no encontrado"
                    },
                    "403": {
                        "description": "No autorizado - Requiere permiso 'Modificar registros'"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Social"
                ],
                "summary": "Actualizar un registro social existente",
                "operationId": "eb462b2b8766f1657ef03ac9586e75c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del registro social a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "integrantes_familia": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro social actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "id_estudiante": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "integrantes_familia": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro no encontrado"
                    },
                    "403": {
                        "description": "No autorizado - Requiere permiso 'Modificar registros'"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Social"
                ],
                "summary": "Eliminar un registro social",
                "operationId": "1625bf47a33cacd7e252029910c9b86a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del registro social a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro social eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro no encontrado"
                    },
                    "403": {
                        "description": "No autorizado - Requiere permiso 'CrearEliminar'"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Sanctum API token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}